Update init data

+++ front/src/services/local.service.js
@@ -1,5 +1,5 @@
 import { saveTemplateAsFile } from '../utils/download/download';
-import { resetData } from '../utils/tree/data';
+import initData from './data.json';
 
 const setItem = (key, value) => {
   window.localStorage.setItem(key, JSON.stringify(value));
@@ -10,15 +10,49 @@ const getItem = (key) => {
 };
 
 const exportStore = () => {
-  saveTemplateAsFile(`data ${new Date().toISOString()}.json`, localStorage);
+  const keys = [
+    'faultTrees',
+    'faultTreeNodes',
+    'faultScenarios',
+    'faultScenarioNodes',
+    'faultTreeNodeDictionary',
+  ];
+
+  const dateStr = new Date().toISOString();
+  saveTemplateAsFile(
+    `data_${dateStr}.json`,
+    keys.reduce(
+      (acc, key) => {
+        acc[key] = getItem(key);
+        return acc;
+      },
+      {
+        dataVersion: dateStr,
+      }
+    )
+  );
 };
+
+export function reloadData(clear, newData) {
+  const data = newData || initData;
+
+  if (clear) {
+    window.localStorage.clear();
+  }
+
+  if (!localStorage.getItem('dataVersion')) {
+    Object.keys(data).forEach((key) => {
+      setItem(key, data[key]);
+    });
+    window.location.reload();
+  }
+}
 
 const reset = () => {
-  window.localStorage.clear();
-  resetData();
+  reloadData(true);
 };
 
-resetData();
+reloadData();
 
 export const localService = {
   setItem,
